Commit "experiment" to adding encoding by using 4 type parameters
The problem with modeling `Input -> T -> Output` decoding + encoding operations with 3 type parameters is that `T` needs to be both covariant and contravariant. Splitting `T` into `T1` and `T2`, at least for type parameters internal to the library, solves this.

Committing on a separate branch as the plan right now is to "give up" an use `any` for the tutorial which can allow `T` to simultaneously be a subtype and supertype of every other type, satisfying type constraints by turning off the type checker.